+++ module/eco/mapper/delegation.go
@@ -1,37 +1,11 @@
 package mapper
 
-import(
-	"encoding/binary"
+import (
+	"github.com/QOSGroup/qbase/context"
 	"github.com/QOSGroup/qbase/mapper"
 	btypes "github.com/QOSGroup/qbase/types"
-	"github.com/QOSGroup/qbase/context"
 	stake_types "github.com/QOSGroup/qos/module/eco/types"
 )
-
-const (
-
-)
-
-var (
-)
-
-func BuildDelegationByDelValKey(delAdd btypes.Address, valAdd btypes.Address) []byte{
-	bz := append(stake_types.DelegationByDelValKey, delAdd...)
-	return append(bz, valAdd...)
-}
-
-func BuildDelegationByValDelKey(valAdd btypes.Address, delAdd btypes.Address) []byte{
-	bz := append(stake_types.DelegationByDelValKey, valAdd...)
-	return append(bz, delAdd...)
-}
-
-func BuildUnbondingDelegationByHeightDelKey(height uint64, delAdd btypes.Address) []byte{
-	heightBytes := make([]byte, 8)
-	binary.BigEndian.PutUint64(heightBytes, height)
-
-	bz := append(stake_types.DelegatorUnbondingQOSatHeightKey, heightBytes...)
-	return append(bz, delAdd...)
-}
 
 type DelegationMapper struct {
 	*mapper.BaseMapper
@@ -50,18 +24,18 @@ func (mapper *DelegationMapper) Copy() mapper.IMapper {
 }
 
 func (mapper *DelegationMapper) SetDelegationInfo(info stake_types.DelegationInfo){
-	mapper.Set(BuildDelegationByDelValKey(info.DelegatorAddr, info.ValidatorAddr), info)
-	mapper.Set(BuildDelegationByValDelKey(info.ValidatorAddr, info.DelegatorAddr), nil)
+	mapper.Set(stake_types.BuildDelegationByDelValKey(info.DelegatorAddr, info.ValidatorAddr), info)
+	mapper.Set(stake_types.BuildDelegationByValDelKey(info.ValidatorAddr, info.DelegatorAddr), nil)
 }
 
 func (mapper *DelegationMapper) GetDelegationInfo(delAddr btypes.Address, valAddr btypes.Address) (info stake_types.DelegationInfo, exist bool) {
-	exist = mapper.Get(BuildDelegationByDelValKey(delAddr, valAddr), &info)
+	exist = mapper.Get(stake_types.BuildDelegationByDelValKey(delAddr, valAddr), &info)
 	return
 }
 
 func (mapper *DelegationMapper) DelDelegationInfo(delAddr btypes.Address, valAddr btypes.Address){
-	mapper.Del(BuildDelegationByDelValKey(delAddr, valAddr))
-	mapper.Del(BuildDelegationByValDelKey(valAddr, delAddr))
+	mapper.Del(stake_types.BuildDelegationByDelValKey(delAddr, valAddr))
+	mapper.Del(stake_types.BuildDelegationByValDelKey(valAddr, delAddr))
 }
 
 func (mapper *DelegationMapper) AddQOStoDelegationInfo(delAddr btypes.Address, valAddr btypes.Address, add_amount uint64) (amount uint64, exist bool) {
@@ -102,11 +76,11 @@ func (mapper *DelegationMapper) ChangeDelegationInfoCompound(delAddr btypes.Addr
 }
 
 func (mapper *DelegationMapper) SetDelegatorUnbondingQOSatHeight(height uint64, delAddr btypes.Address, amount uint64){
-	mapper.Set(BuildUnbondingDelegationByHeightDelKey(height, delAddr), amount)
+	mapper.Set(stake_types.BuildUnbondingDelegationByHeightDelKey(height, delAddr), amount)
 }
 
 func (mapper *DelegationMapper) GetDelegatorUnbondingQOSatHeight(height uint64, delAdd btypes.Address) (amount uint64, exist bool){
-	exist = mapper.Get(BuildUnbondingDelegationByHeightDelKey(height, delAdd), &amount)
+	exist = mapper.Get(stake_types.BuildUnbondingDelegationByHeightDelKey(height, delAdd), &amount)
 	return
 }
 
@@ -119,7 +93,7 @@ func (mapper *DelegationMapper) AddDelegatorUnbondingQOSatHeight(height uint64, 
 }
 
 func (mapper *DelegationMapper) RemoveDelegatorUnbondingQOSatHeight(height uint64, delAddr btypes.Address) {
-	mapper.Del(BuildUnbondingDelegationByHeightDelKey(height, delAddr))
+	mapper.Del(stake_types.BuildUnbondingDelegationByHeightDelKey(height, delAddr))
 }
 
 func (mapper *DelegationMapper) CreateDelegation(delAddr btypes.Address, valAddr btypes.Address, amount uint64, isCompound bool) (info stake_types.DelegationInfo){
